refactor(ui.windows.line-numbers): Replace `gitgutter` with `gitsigns`
Integrates with `barbar` (the tab bar plugin)... Well they say that but
I saw no difference. I think `gitsigns` looks better anyway and I like
its features more. `gitgutter` was not supposed to stay, but it was here
because `gitsigns` wasn't working—turns out it was because `gitgutter`
was enabled lol.