chore(ui.tab-bar.keys): Prefer `<leader>` keystrokes over `<C-b>` shortcuts
This is mainly due to the fact that `<C-b>` is a keymap to scroll down
the cursor, and I couldn't find a way to unmap it.
There is also an argument for consistency but it falls appart when you
consider window manipulation shortcuts.
Vim shortcuts can be weird sometimes, huh... (at least it's not Emacs')